update redbus scapring python file
Previous version of scraping code will have rough process of code. with this new version updated with try except for error handing selenium eevents, all the formate & its element are added in a dictonary of classification. each module is break up into function so that continous scapring will be process.